Update mira2_image.i
Just to have also in the header the GPNORM value to better know from the header file if the True convergence is robust or not, It happened many times that the convergence has been set True while GPNORM reached still very high values. I think it can be a useful information to get.